refactor(node): split client core transfers
Extract transfer operations and policy handling into dedicated helpers to reduce NodeClientCore dependencies.
Update call sites and tests to use getTransfers() access.
